Draft: decouple kubeseal api
