+++ pkg/kubeseal/kubeseal.go
@@ -13,15 +13,12 @@ import (
 	"net/http"
 	"net/url"
 	"os"
-	"path/filepath"
 	"strings"
 	"time"
 
 	ssv1alpha1 "github.com/bitnami-labs/sealed-secrets/pkg/apis/sealedsecrets/v1alpha1"
 	"github.com/bitnami-labs/sealed-secrets/pkg/crypto"
 	"github.com/bitnami-labs/sealed-secrets/pkg/multidocyaml"
-	"github.com/google/renameio"
-	"github.com/mattn/go-isatty"
 	v1 "k8s.io/api/core/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -35,57 +32,9 @@ import (
 	"k8s.io/client-go/util/keyutil"
 )
 
-type namespaceFn func() (string, bool, error)
+type NamespaceFn func() (string, bool, error)
 
-type Flags struct {
-	CertURL        string
-	ControllerNs   string
-	ControllerName string
-	OutputFormat   string
-	OutputFileName string
-	InputFileName  string
-	Kubeconfig     string
-	DumpCert       bool
-	AllowEmptyData bool
-	ValidateSecret bool
-	MergeInto      string
-	Raw            bool
-	SecretName     string
-	FromFile       []string
-	SealingScope   ssv1alpha1.SealingScope
-	ReEncrypt      bool
-	Unseal         bool
-	PrivKeys       []string
-}
-
-type Config struct {
-	flags          *Flags
-	clientConfig   clientcmd.ClientConfig
-	ctx            context.Context
-	solveNamespace namespaceFn
-}
-
-func initNamespaceFuncFromClient(clientConfig clientcmd.ClientConfig) namespaceFn {
-	return func() (string, bool, error) { return clientConfig.Namespace() }
-}
-
-func NewConfig(clientConfig clientcmd.ClientConfig, flags *Flags) *Config {
-	return &Config{
-		flags:          flags,
-		clientConfig:   clientConfig,
-		ctx:            context.Background(),
-		solveNamespace: initNamespaceFuncFromClient(clientConfig),
-	}
-}
-
-func InitClient(kubeConfigPath string, cfgOverrides *clientcmd.ConfigOverrides, r io.Reader) clientcmd.ClientConfig {
-	loadingRules := clientcmd.NewDefaultClientConfigLoadingRules()
-	loadingRules.DefaultClientConfig = &clientcmd.DefaultClientConfig
-	loadingRules.ExplicitPath = kubeConfigPath
-	return clientcmd.NewInteractiveDeferredLoadingClientConfig(loadingRules, cfgOverrides, r)
-}
-
-func parseKey(r io.Reader) (*rsa.PublicKey, error) {
+func ParseKey(r io.Reader) (*rsa.PublicKey, error) {
 	data, err := io.ReadAll(r)
 	if err != nil {
 		return nil, err
@@ -214,12 +163,12 @@ func openCertCluster(ctx context.Context, c corev1.CoreV1Interface, namespace, n
 	return cert, nil
 }
 
-func openCert(cfg *Config, certURL string) (io.ReadCloser, error) {
+func OpenCert(ctx context.Context, clientConfig clientcmd.ClientConfig, controllerNs, controllerName string, certURL string) (io.ReadCloser, error) {
 	if certURL != "" {
 		return openCertLocal(certURL)
 	}
 
-	conf, err := cfg.clientConfig.ClientConfig()
+	conf, err := clientConfig.ClientConfig()
 	if err != nil {
 		return nil, err
 	}
@@ -228,13 +177,13 @@ func openCert(cfg *Config, certURL string) (io.ReadCloser, error) {
 	if err != nil {
 		return nil, err
 	}
-	return openCertCluster(cfg.ctx, restClient, cfg.flags.ControllerNs, cfg.flags.ControllerName)
+	return openCertCluster(ctx, restClient, controllerNs, controllerName)
 }
 
 // Seal reads a k8s Secret resource parsed from an input reader by a given codec, encrypts all its secrets
 // with a given public key, using the name and namespace found in the input secret, unless explicitly overridden
 // by the overrideName and overrideNamespace arguments.
-func seal(cfg *Config, in io.Reader, out io.Writer, codecs runtimeserializer.CodecFactory, pubKey *rsa.PublicKey, scope ssv1alpha1.SealingScope, allowEmptyData bool, overrideName, overrideNamespace string) error {
+func Seal(solveNamespace NamespaceFn, outputFormat string, in io.Reader, out io.Writer, codecs runtimeserializer.CodecFactory, pubKey *rsa.PublicKey, scope ssv1alpha1.SealingScope, allowEmptyData bool, overrideName, overrideNamespace string) error {
 	secret, err := readSecret(codecs.UniversalDecoder(), in)
 	if err != nil {
 		return err
@@ -261,7 +210,7 @@ func seal(cfg *Config, in io.Reader, out io.Writer, codecs runtimeserializer.Cod
 	}
 
 	if ssv1alpha1.SecretScope(secret) != ssv1alpha1.ClusterWideScope && secret.GetNamespace() == "" {
-		ns, _, err := cfg.solveNamespace()
+		ns, _, err := solveNamespace()
 		if clientcmd.IsEmptyConfig(err) {
 			return fmt.Errorf("input secret has no namespace and cannot infer the namespace automatically when no kube config is available")
 		} else if err != nil {
@@ -283,15 +232,14 @@ func seal(cfg *Config, in io.Reader, out io.Writer, codecs runtimeserializer.Cod
 	if err != nil {
 		return err
 	}
-	if err = sealedSecretOutput(out, cfg.flags, codecs, ssecret); err != nil {
+	if err = sealedSecretOutput(out, outputFormat, codecs, ssecret); err != nil {
 		return err
 	}
 	return nil
 }
 
-func validateSealedSecret(cfg *Config, in io.Reader) error {
-	flags := cfg.flags
-	conf, err := cfg.clientConfig.ClientConfig()
+func ValidateSealedSecret(ctx context.Context, clientConfig clientcmd.ClientConfig, controllerNs, controllerName string, in io.Reader) error {
+	conf, err := clientConfig.ClientConfig()
 	if err != nil {
 		return err
 	}
@@ -299,7 +247,7 @@ func validateSealedSecret(cfg *Config, in io.Reader) error {
 	if err != nil {
 		return err
 	}
-	portName, err := getServicePortName(cfg.ctx, restClient, flags.ControllerNs, flags.ControllerName)
+	portName, err := getServicePortName(ctx, restClient, controllerNs, controllerName)
 	if err != nil {
 		return err
 	}
@@ -310,14 +258,14 @@ func validateSealedSecret(cfg *Config, in io.Reader) error {
 	}
 
 	req := restClient.RESTClient().Post().
-		Namespace(flags.ControllerNs).
+		Namespace(controllerNs).
 		Resource("services").
 		SubResource("proxy").
-		Name(net.JoinSchemeNamePort("http", flags.ControllerName, portName)).
+		Name(net.JoinSchemeNamePort("http", controllerName, portName)).
 		Suffix("/v1/verify")
 
 	req.Body(content)
-	res := req.Do(cfg.ctx)
+	res := req.Do(ctx)
 	if err := res.Error(); err != nil {
 		if status, ok := err.(*k8serrors.StatusError); ok && status.Status().Code == http.StatusConflict {
 			return fmt.Errorf("unable to decrypt sealed secret")
@@ -328,9 +276,8 @@ func validateSealedSecret(cfg *Config, in io.Reader) error {
 	return nil
 }
 
-func reEncryptSealedSecret(cfg *Config, in io.Reader, out io.Writer, codecs runtimeserializer.CodecFactory) error {
-	flags := cfg.flags
-	conf, err := cfg.clientConfig.ClientConfig()
+func ReEncryptSealedSecret(ctx context.Context, clientConfig clientcmd.ClientConfig, controllerNs, controllerName, outputFormat string, in io.Reader, out io.Writer, codecs runtimeserializer.CodecFactory) error {
+	conf, err := clientConfig.ClientConfig()
 	if err != nil {
 		return err
 	}
@@ -338,7 +285,7 @@ func reEncryptSealedSecret(cfg *Config, in io.Reader, out io.Writer, codecs runt
 	if err != nil {
 		return err
 	}
-	portName, err := getServicePortName(cfg.ctx, restClient, flags.ControllerNs, flags.ControllerName)
+	portName, err := getServicePortName(ctx, restClient, controllerNs, controllerName)
 	if err != nil {
 		return err
 	}
@@ -349,14 +296,14 @@ func reEncryptSealedSecret(cfg *Config, in io.Reader, out io.Writer, codecs runt
 	}
 
 	req := restClient.RESTClient().Post().
-		Namespace(flags.ControllerNs).
+		Namespace(controllerNs).
 		Resource("services").
 		SubResource("proxy").
-		Name(net.JoinSchemeNamePort("http", flags.ControllerName, portName)).
+		Name(net.JoinSchemeNamePort("http", controllerName, portName)).
 		Suffix("/v1/rotate")
 
 	req.Body(content)
-	res := req.Do(cfg.ctx)
+	res := req.Do(ctx)
 	if err := res.Error(); err != nil {
 		if status, ok := err.(*k8serrors.StatusError); ok && status.Status().Code == http.StatusConflict {
 			return fmt.Errorf("unable to rotate secret")
@@ -374,21 +321,21 @@ func reEncryptSealedSecret(cfg *Config, in io.Reader, out io.Writer, codecs runt
 	ssecret.SetCreationTimestamp(metav1.Time{})
 	ssecret.SetDeletionTimestamp(nil)
 	ssecret.Generation = 0
-	if err = sealedSecretOutput(out, cfg.flags, codecs, ssecret); err != nil {
+	if err = sealedSecretOutput(out, outputFormat, codecs, ssecret); err != nil {
 		return err
 	}
 	return nil
 }
 
-func resourceOutput(out io.Writer, flags *Flags, codecs runtimeserializer.CodecFactory, gv runtime.GroupVersioner, obj runtime.Object) error {
+func resourceOutput(out io.Writer, outputFormat string, codecs runtimeserializer.CodecFactory, gv runtime.GroupVersioner, obj runtime.Object) error {
 	var contentType string
-	switch strings.ToLower(flags.OutputFormat) {
+	switch strings.ToLower(outputFormat) {
 	case "json", "":
 		contentType = runtime.ContentTypeJSON
 	case "yaml":
 		contentType = runtime.ContentTypeYAML
 	default:
-		return fmt.Errorf("unsupported output format: %s", flags.OutputFormat)
+		return fmt.Errorf("unsupported output format: %s", outputFormat)
 	}
 	prettyEnc, err := prettyEncoder(codecs, contentType, gv)
 	if err != nil {
@@ -403,8 +350,8 @@ func resourceOutput(out io.Writer, flags *Flags, codecs runtimeserializer.CodecF
 	return nil
 }
 
-func sealedSecretOutput(out io.Writer, flags *Flags, codecs runtimeserializer.CodecFactory, ssecret *ssv1alpha1.SealedSecret) error {
-	return resourceOutput(out, flags, codecs, ssv1alpha1.SchemeGroupVersion, ssecret)
+func sealedSecretOutput(out io.Writer, outputFormat string, codecs runtimeserializer.CodecFactory, ssecret *ssv1alpha1.SealedSecret) error {
+	return resourceOutput(out, outputFormat, codecs, ssv1alpha1.SchemeGroupVersion, ssecret)
 }
 
 func decodeSealedSecret(codecs runtimeserializer.CodecFactory, b []byte) (*ssv1alpha1.SealedSecret, error) {
@@ -415,7 +362,7 @@ func decodeSealedSecret(codecs runtimeserializer.CodecFactory, b []byte) (*ssv1a
 	return &ss, nil
 }
 
-func sealMergingInto(cfg *Config, in io.Reader, filename string, codecs runtimeserializer.CodecFactory, pubKey *rsa.PublicKey, scope ssv1alpha1.SealingScope, allowEmptyData bool) error {
+func SealMergingInto(solveNamespace NamespaceFn, outputFormat string, in io.Reader, filename string, codecs runtimeserializer.CodecFactory, pubKey *rsa.PublicKey, scope ssv1alpha1.SealingScope, allowEmptyData bool) error {
 	// #nosec G304 -- should open user provided file
 	f, err := os.OpenFile(filename, os.O_RDWR, 0)
 	if err != nil {
@@ -435,7 +382,7 @@ func sealMergingInto(cfg *Config, in io.Reader, filename string, codecs runtimes
 	}
 
 	var buf bytes.Buffer
-	if err := seal(cfg, in, &buf, codecs, pubKey, scope, allowEmptyData, orig.Name, orig.Namespace); err != nil {
+	if err := Seal(solveNamespace, outputFormat, in, &buf, codecs, pubKey, scope, allowEmptyData, orig.Name, orig.Namespace); err != nil {
 		return err
 	}
 
@@ -460,7 +407,7 @@ func sealMergingInto(cfg *Config, in io.Reader, filename string, codecs runtimes
 
 	// updated sealed secret file in-place avoiding clobbering the file upon rendering errors.
 	var out bytes.Buffer
-	if err := sealedSecretOutput(&out, cfg.flags, codecs, orig); err != nil {
+	if err := sealedSecretOutput(&out, outputFormat, codecs, orig); err != nil {
 		return err
 	}
 
@@ -480,7 +427,7 @@ func sealMergingInto(cfg *Config, in io.Reader, filename string, codecs runtimes
 	return nil
 }
 
-func encryptSecretItem(w io.Writer, secretName, ns string, data []byte, scope ssv1alpha1.SealingScope, pubKey *rsa.PublicKey) error {
+func EncryptSecretItem(w io.Writer, secretName, ns string, data []byte, scope ssv1alpha1.SealingScope, pubKey *rsa.PublicKey) error {
 	// TODO(mkm): refactor cluster-wide/namespace-wide to an actual enum so we can have a simple flag
 	// to refer to the scope mode that is not a tuple of booleans.
 	label := ssv1alpha1.EncryptionLabel(ns, secretName, scope)
@@ -494,7 +441,7 @@ func encryptSecretItem(w io.Writer, secretName, ns string, data []byte, scope ss
 
 // parseFromFile parses a value of the kubectl --from-file flag, which can optionally include an item name
 // preceding the first equals sign.
-func parseFromFile(s string) (string, string) {
+func ParseFromFile(s string) (string, string) {
 	c := strings.SplitN(s, "=", 2)
 	if len(c) == 1 {
 		return "", c[0]
@@ -591,8 +538,8 @@ func readPrivKeys(filenames []string) (map[string]*rsa.PrivateKey, error) {
 	return res, nil
 }
 
-func unsealSealedSecret(flags *Flags, w io.Writer, in io.Reader, codecs runtimeserializer.CodecFactory) error {
-	privKeys, err := readPrivKeys(flags.PrivKeys)
+func UnsealSealedSecret(w io.Writer, in io.Reader, privKeysFilenames []string, outputFormat string, codecs runtimeserializer.CodecFactory) error {
+	privKeys, err := readPrivKeys(privKeysFilenames)
 	if err != nil {
 		return err
 	}
@@ -611,133 +558,5 @@ func unsealSealedSecret(flags *Flags, w io.Writer, in io.Reader, codecs runtimes
 		return err
 	}
 
-	return resourceOutput(w, flags, codecs, v1.SchemeGroupVersion, sec)
-}
-
-func Run(w io.Writer, cfg *Config) (err error) {
-	flags := cfg.flags
-	if len(flags.FromFile) != 0 && !flags.Raw {
-		return fmt.Errorf("--from-file requires --raw")
-	}
-
-	var input io.Reader = os.Stdin
-	if flags.InputFileName != "" {
-		// #nosec G304 -- should open user provided file
-		f, err := os.Open(flags.InputFileName)
-		if err != nil {
-			return nil
-		}
-		// #nosec: G307 -- this deferred close is fine because it is not on a writable file
-		defer f.Close()
-
-		input = f
-	} else if !flags.Raw && !flags.DumpCert {
-		if isatty.IsTerminal(os.Stdin.Fd()) {
-			fmt.Fprintf(os.Stderr, "(tty detected: expecting json/yaml k8s resource in stdin)\n")
-		}
-	}
-
-	// reEncrypt is the only "in-place" update subcommand. When the user only provides one file (the input file)
-	// we'll use the same file for output (see #405).
-	if flags.ReEncrypt && (flags.OutputFileName == "" && flags.InputFileName != "") {
-		flags.OutputFileName = flags.InputFileName
-	}
-	if flags.OutputFileName != "" {
-		if ext := filepath.Ext(flags.OutputFileName); ext == ".yaml" || ext == ".yml" {
-			flags.OutputFormat = "yaml"
-		}
-
-		var f *renameio.PendingFile
-		f, err = renameio.TempFile("", flags.OutputFileName)
-		if err != nil {
-			return err
-		}
-		// only write the output file if the run function exits without errors.
-		defer func() {
-			if err == nil {
-				_ = f.CloseAtomicallyReplace()
-			}
-		}()
-
-		w = f
-	}
-
-	if flags.Unseal {
-		return unsealSealedSecret(flags, w, input, scheme.Codecs)
-	}
-	if len(flags.PrivKeys) != 0 && isatty.IsTerminal(os.Stderr.Fd()) {
-		fmt.Fprintf(os.Stderr, "warning: ignoring --recovery-private-key because unseal command not chosen with --recovery-unseal\n")
-	}
-
-	if flags.ValidateSecret {
-		return validateSealedSecret(cfg, input)
-	}
-
-	if flags.ReEncrypt {
-		return reEncryptSealedSecret(cfg, input, w, scheme.Codecs)
-	}
-
-	f, err := openCert(cfg, flags.CertURL)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	if flags.DumpCert {
-		_, err := io.Copy(w, f)
-		return err
-	}
-
-	pubKey, err := parseKey(f)
-	if err != nil {
-		return err
-	}
-
-	if flags.MergeInto != "" {
-		return sealMergingInto(cfg, input, flags.MergeInto, scheme.Codecs, pubKey, flags.SealingScope, flags.AllowEmptyData)
-	}
-
-	if flags.Raw {
-		var (
-			ns  string
-			err error
-		)
-		if flags.SealingScope < ssv1alpha1.ClusterWideScope {
-			ns, _, err = cfg.solveNamespace()
-			if err != nil {
-				return err
-			}
-
-			if ns == "" {
-				return fmt.Errorf("must provide the --namespace flag with --raw and --scope %s", flags.SealingScope.String())
-			}
-
-			if flags.SecretName == "" && flags.SealingScope < ssv1alpha1.NamespaceWideScope {
-				return fmt.Errorf("must provide the --name flag with --raw and --scope %s", flags.SealingScope.String())
-			}
-		}
-
-		var data []byte
-		if len(flags.FromFile) > 0 {
-			if len(flags.FromFile) > 1 {
-				return fmt.Errorf("must provide only one --from-file when encrypting a single item with --raw")
-			}
-
-			_, filename := parseFromFile(flags.FromFile[0])
-			// #nosec G304 -- should open user provided file
-			data, err = os.ReadFile(filename)
-		} else {
-			if isatty.IsTerminal(os.Stdin.Fd()) {
-				fmt.Fprintf(os.Stderr, "(tty detected: expecting a secret to encrypt in stdin)\n")
-			}
-			data, err = io.ReadAll(os.Stdin)
-		}
-		if err != nil {
-			return err
-		}
-
-		return encryptSecretItem(w, flags.SecretName, ns, data, flags.SealingScope, pubKey)
-	}
-
-	return seal(cfg, input, w, scheme.Codecs, pubKey, flags.SealingScope, flags.AllowEmptyData, flags.SecretName, "")
+	return resourceOutput(w, outputFormat, codecs, v1.SchemeGroupVersion, sec)
 }
